Move devices in subfolder
