+++ components/gpio_control/gpio_control.py
@@ -2,11 +2,12 @@ import configparser
 import os
 import logging
 
-from .simple_button import SimpleButton
-from .two_button_control import TwoButtonControl
+from .GPIODevices.simple_button import SimpleButton
+from .GPIODevices.two_button_control import TwoButtonControl
+from .GPIODevices.rotary_encoder import RotaryEncoder
 from . import  function_calls
 from signal import pause
-from .rotary_encoder import RotaryEncoder
+
 from RPi import GPIO
 
 GPIO.setmode(GPIO.BCM)
